+++ src/traits/mod.rs
@@ -4,8 +4,13 @@
 
 mod btoi;
 mod btou;
+mod inflect;
 mod saturating_from;
 
 pub use btoi::BytesToSigned;
 pub use btou::BytesToUnsigned;
+pub use inflect::{
+	Inflection,
+	NiceInflection,
+};
 pub use saturating_from::SaturatingFrom;